+++ defaults/before-build.js
@@ -1,4 +1,4 @@
 module.exports = {
     keyFileProjectConfigTemp: '__FILE_PROJECT_CONFIG_TEMP__',
-    filenameProjectConfigTemp: '.koot.config.tmp.[TS].js'
+    filenameProjectConfigTemp: '.koot.config.tmp.*.js'
 }